+++ lib/active_admin/page_dsl.rb
@@ -25,8 +25,9 @@ module ActiveAdmin
       end
     end
 
-    def belongs_to(target, options = {})
-      config.belongs_to(target, options)
+    def belongs_to(*targets)
+      options = targets.extract_options!
+      config.belongs_to(targets, options)
     end
   end
 end