+++ spec/unit/namespace/register_resource_spec.rb
@@ -127,10 +127,22 @@ RSpec.describe ActiveAdmin::Namespace, "registering a resource" do
           expect(menu["Posts"]).to eq nil
         end
       end
+
       context "when optional" do
         before do
           namespace.register Post do
             belongs_to :author, optional: true
+          end
+        end
+        it "should show up in the menu" do
+          expect(menu["Posts"]).to_not eq nil
+        end
+      end
+
+      context "with multiple optional parents" do
+        before do
+          namespace.register Post do
+            belongs_to :author, :user, optional: true
           end
         end
         it "should show up in the menu" do